+++ src/main/java/io/secuconnect/client/api/SmartTransactionsApi.java
@@ -1154,13 +1154,14 @@ public class SmartTransactionsApi {
      * Build call for startTransaction
      * @param smartTransactionId Smart transaction id (required)
      * @param paymentMethod Payment method (required)
+     * @param body Information which customer and container will be used to this operation
      * @param progressListener Progress listener
      * @param progressRequestListener Progress request listener
      * @return Call to execute
      * @throws ApiException If fail to serialize the request body object
      */
-    public com.squareup.okhttp.Call startTransactionCall(String smartTransactionId, String paymentMethod, final ProgressResponseBody.ProgressListener progressListener, final ProgressRequestBody.ProgressRequestListener progressRequestListener) throws ApiException {
-        Object localVarPostBody = null;
+    public com.squareup.okhttp.Call startTransactionCall(String smartTransactionId, String paymentMethod, SmartTransactionsPrepare body, final ProgressResponseBody.ProgressListener progressListener, final ProgressRequestBody.ProgressRequestListener progressRequestListener) throws ApiException {
+        Object localVarPostBody = body;
         
         // create path and map variables
         String localVarPath = "/Smart/Transactions/{smartTransactionId}/start/{paymentMethod}"
@@ -1181,7 +1182,7 @@ public class SmartTransactionsApi {
         if (localVarAccept != null) localVarHeaderParams.put("Accept", localVarAccept);
 
         final String[] localVarContentTypes = {
-            
+            "application/json"
         };
         final String localVarContentType = apiClient.selectHeaderContentType(localVarContentTypes);
         localVarHeaderParams.put("Content-Type", localVarContentType);
@@ -1203,7 +1204,7 @@ public class SmartTransactionsApi {
     }
     
     @SuppressWarnings("rawtypes")
-    private com.squareup.okhttp.Call startTransactionValidateBeforeCall(String smartTransactionId, String paymentMethod, final ProgressResponseBody.ProgressListener progressListener, final ProgressRequestBody.ProgressRequestListener progressRequestListener) throws ApiException {
+    private com.squareup.okhttp.Call startTransactionValidateBeforeCall(String smartTransactionId, String paymentMethod, SmartTransactionsPrepare body, final ProgressResponseBody.ProgressListener progressListener, final ProgressRequestBody.ProgressRequestListener progressRequestListener) throws ApiException {
         // verify the required parameter 'smartTransactionId' is set
         if (smartTransactionId == null) {
             throw new ApiException("Missing the required parameter 'smartTransactionId' when calling startTransaction(Async)");
@@ -1213,7 +1214,7 @@ public class SmartTransactionsApi {
             throw new ApiException("Missing the required parameter 'paymentMethod' when calling startTransaction(Async)");
         }
         
-        com.squareup.okhttp.Call call = startTransactionCall(smartTransactionId, paymentMethod, progressListener, progressRequestListener);
+        com.squareup.okhttp.Call call = startTransactionCall(smartTransactionId, paymentMethod, body, progressListener, progressRequestListener);
         return call;
     }
 
@@ -1222,11 +1223,12 @@ public class SmartTransactionsApi {
      * Start smart transaction with given payment method
      * @param smartTransactionId Smart transaction id (required)
      * @param paymentMethod Payment method (required)
+     * @param body Information which customer and container will be used to this operation
      * @return SmartTransactionsProductModel
      * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
      */
-    public SmartTransactionsProductModel startTransaction(String smartTransactionId, String paymentMethod) throws ApiException {
-        ApiResponse<SmartTransactionsProductModel> resp = startTransactionWithHttpInfo(smartTransactionId, paymentMethod);
+    public SmartTransactionsProductModel startTransaction(String smartTransactionId, String paymentMethod, SmartTransactionsPrepare body) throws ApiException {
+        ApiResponse<SmartTransactionsProductModel> resp = startTransactionWithHttpInfo(smartTransactionId, paymentMethod, body);
         return resp.getData();
     }
 
@@ -1235,11 +1237,12 @@ public class SmartTransactionsApi {
      * Start smart transaction with given payment method
      * @param smartTransactionId Smart transaction id (required)
      * @param paymentMethod Payment method (required)
+     * @param body Information which customer and container will be used to this operation
      * @return ApiResponse&lt;SmartTransactionsProductModel&gt;
      * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
      */
-    public ApiResponse<SmartTransactionsProductModel> startTransactionWithHttpInfo(String smartTransactionId, String paymentMethod) throws ApiException {
-        com.squareup.okhttp.Call call = startTransactionValidateBeforeCall(smartTransactionId, paymentMethod, null, null);
+    public ApiResponse<SmartTransactionsProductModel> startTransactionWithHttpInfo(String smartTransactionId, String paymentMethod, SmartTransactionsPrepare body) throws ApiException {
+        com.squareup.okhttp.Call call = startTransactionValidateBeforeCall(smartTransactionId, paymentMethod, body, null, null);
         Type localVarReturnType = new TypeToken<SmartTransactionsProductModel>(){}.getType();
         return apiClient.execute(call, localVarReturnType);
     }
@@ -1249,11 +1252,12 @@ public class SmartTransactionsApi {
      * Start smart transaction with given payment method
      * @param smartTransactionId Smart transaction id (required)
      * @param paymentMethod Payment method (required)
+     * @param body Information which customer and container will be used to this operation
      * @param callback The callback to be executed when the API call finishes
      * @return The request call
      * @throws ApiException If fail to process the API call, e.g. serializing the request body object
      */
-    public com.squareup.okhttp.Call startTransactionAsync(String smartTransactionId, String paymentMethod, final ApiCallback<SmartTransactionsProductModel> callback) throws ApiException {
+    public com.squareup.okhttp.Call startTransactionAsync(String smartTransactionId, String paymentMethod, SmartTransactionsPrepare body, final ApiCallback<SmartTransactionsProductModel> callback) throws ApiException {
 
         ProgressResponseBody.ProgressListener progressListener = null;
         ProgressRequestBody.ProgressRequestListener progressRequestListener = null;
@@ -1274,7 +1278,7 @@ public class SmartTransactionsApi {
             };
         }
 
-        com.squareup.okhttp.Call call = startTransactionValidateBeforeCall(smartTransactionId, paymentMethod, progressListener, progressRequestListener);
+        com.squareup.okhttp.Call call = startTransactionValidateBeforeCall(smartTransactionId, paymentMethod, body, progressListener, progressRequestListener);
         Type localVarReturnType = new TypeToken<SmartTransactionsProductModel>(){}.getType();
         apiClient.executeAsync(call, localVarReturnType, callback);
         return call;